fix stats model

+++ src/Saiive.SuperNode.Model/StatsModel.cs
@@ -42,7 +42,7 @@ namespace Saiive.SuperNode.Model
         public double Emission { get; set; }
 
         [JsonProperty("fee")]
-        public int Fee { get; set; }
+        public double Fee { get; set; }
 
         [JsonProperty("total")]
         public double Total { get; set; }
